chore: revert deployment scripts
